spannertadmin/requirements.txt
Django==4.1.3
djangorestframework==3.14.0
mysqlclient==2.1.1
django-mysql==4.7.1
django-cors-headers==3.13.0
pika==1.3.1
psycopg2-binary==2.9.1